refactor(wow-core): move wait strategy constants and logic to WaitingForStage
- Move COMMAND_WAIT constants to WaitingForStage companion object
- Extract injectWaitStrategy logic into WaitingForStage
- Update related tests and classes to use new WaitingForStage.inject method
- Remove redundant injectWaitStrategy function from Header
