+++ wow-core/src/test/kotlin/me/ahoo/wow/command/wait/MonoCommandWaitNotifierTest.kt
@@ -43,8 +43,7 @@ internal class MonoCommandWaitNotifierTest {
     @Test
     fun notifyAndForgetWrap() {
         val command = MockCreateCommand("").toCommandMessage()
-
-        command.header.injectWaitStrategy("", WaitingForStage.processed())
+        WaitingForStage.processed().inject(SimpleCommandWaitEndpoint(""), command.header)
 
         val commandWaitNotifier = LocalCommandWaitNotifier(SimpleWaitStrategyRegistrar)
         Mono.empty<Void>()
@@ -65,7 +64,7 @@ internal class MonoCommandWaitNotifierTest {
     @Test
     fun notifyAndForgetWrapError() {
         val command = MockCreateCommand("").toCommandMessage()
-        command.header.injectWaitStrategy("", WaitingForStage.processed())
+        WaitingForStage.processed().inject(SimpleCommandWaitEndpoint(""), command.header)
         val commandWaitNotifier = LocalCommandWaitNotifier(SimpleWaitStrategyRegistrar)
         RuntimeException("error")
             .toMono<Void>()
@@ -81,7 +80,7 @@ internal class MonoCommandWaitNotifierTest {
     @Test
     fun notifyAndForgetWrapAndStageIsEarly() {
         val command = MockCreateCommand("").toCommandMessage()
-        command.header.injectWaitStrategy("", WaitingForStage.processed())
+        WaitingForStage.processed().inject(SimpleCommandWaitEndpoint(""), command.header)
         val commandWaitNotifier = LocalCommandWaitNotifier(SimpleWaitStrategyRegistrar)
         Mono.empty<Void>()
             .thenNotifyAndForget(commandWaitNotifier, CommandStage.PROCESSED, SimpleServerCommandExchange(command))
