+++ wow-core/src/test/kotlin/me/ahoo/wow/messaging/propagation/WaitStrategyMessagePropagatorTest.kt
@@ -2,14 +2,14 @@ package me.ahoo.wow.messaging.propagation
 
 import me.ahoo.test.asserts.assert
 import me.ahoo.wow.command.toCommandMessage
-import me.ahoo.wow.command.wait.COMMAND_WAIT_CONTEXT
 import me.ahoo.wow.command.wait.COMMAND_WAIT_ENDPOINT
-import me.ahoo.wow.command.wait.COMMAND_WAIT_FUNCTION
-import me.ahoo.wow.command.wait.COMMAND_WAIT_PROCESSOR
-import me.ahoo.wow.command.wait.COMMAND_WAIT_STAGE
-import me.ahoo.wow.command.wait.injectWaitStrategy
+import me.ahoo.wow.command.wait.SimpleCommandWaitEndpoint
 import me.ahoo.wow.command.wait.stage.WaitingForStage
-import me.ahoo.wow.id.GlobalIdGenerator
+import me.ahoo.wow.command.wait.stage.WaitingForStage.Companion.COMMAND_WAIT_CONTEXT
+import me.ahoo.wow.command.wait.stage.WaitingForStage.Companion.COMMAND_WAIT_FUNCTION
+import me.ahoo.wow.command.wait.stage.WaitingForStage.Companion.COMMAND_WAIT_PROCESSOR
+import me.ahoo.wow.command.wait.stage.WaitingForStage.Companion.COMMAND_WAIT_STAGE
+import me.ahoo.wow.id.generateGlobalId
 import me.ahoo.wow.messaging.DefaultHeader
 import me.ahoo.wow.tck.mock.MockCreateAggregate
 import org.junit.jupiter.api.Test
@@ -20,13 +20,9 @@ class WaitStrategyMessagePropagatorTest {
     fun inject() {
         val header = DefaultHeader.empty()
         val upstreamMessage =
-            MockCreateAggregate(GlobalIdGenerator.generateAsString(), GlobalIdGenerator.generateAsString())
+            MockCreateAggregate(generateGlobalId(), generateGlobalId())
                 .toCommandMessage()
-
-        upstreamMessage.header.injectWaitStrategy(
-            "wait-endpoint",
-            WaitingForStage.projected("context", "processor", "function")
-        )
+        WaitingForStage.processed().inject(SimpleCommandWaitEndpoint("wait-endpoint"), upstreamMessage.header)
         WaitStrategyMessagePropagator().inject(header, upstreamMessage)
         header[COMMAND_WAIT_ENDPOINT].assert().isEqualTo("wait-endpoint")
         header[COMMAND_WAIT_STAGE].assert().isEqualTo("PROJECTED")
@@ -39,9 +35,9 @@ class WaitStrategyMessagePropagatorTest {
     fun injectIfBlank() {
         val header = DefaultHeader.empty()
         val upstreamMessage =
-            MockCreateAggregate(GlobalIdGenerator.generateAsString(), GlobalIdGenerator.generateAsString())
+            MockCreateAggregate(generateGlobalId(), generateGlobalId())
                 .toCommandMessage()
-        upstreamMessage.header.injectWaitStrategy("wait-endpoint", WaitingForStage.sent())
+        WaitingForStage.sent().inject(SimpleCommandWaitEndpoint("wait-endpoint"), upstreamMessage.header)
         WaitStrategyMessagePropagator().inject(header, upstreamMessage)
         header[COMMAND_WAIT_ENDPOINT].assert().isEqualTo(upstreamMessage.header[COMMAND_WAIT_ENDPOINT])
         header[COMMAND_WAIT_STAGE].assert().isEqualTo(upstreamMessage.header[COMMAND_WAIT_STAGE])
